refactor: move location queries to separate module and improve validation
