+++ pages/dashboard/add.vue
@@ -31,7 +31,7 @@ const onSubmit = handleSubmit(async (values) => {
     if (error.data?.data) {
       setErrors(error.data?.data);
     }
-    submitError.value = error.statusMessage || "An unknown error occurred";
+    submitError.value = error.data?.statusMessage || error.statusMessage || "An unknown error occurred";
   }
   loading.value = false;
 });
